Split handlers.ts phase 3 — extract remaining handlers into per-domain files (warren-599c / pl-9088 step 3)
src/server/handlers/index.ts (1194 lines) splits into per-domain handler
modules following the phase-1/phase-2 pattern (mx-3df5c5, mx-99ad0d):

- agents.ts (203) — list/get/refresh, plus POST /projects/:id/agents/refresh
- projects.ts (261) — list/create/delete + warren-config/triggers/seeds/refresh
- burrows.ts (116) — multi-worker list + sticky get
- workers.ts (89) — list + drain admin
- diagnostics.ts (136) — /readyz + readyz sub-probes
- meta.ts (42) — /healthz + /version + /preview/config

handlers/index.ts shrinks to 420 lines and now contains only the
ROUTE_TABLE composer, shared parsing helpers (readJsonBody,
requireString, etc., defaultSpawn, assertPlotIdDispatchable), and
API_PREFIXES / isAuthExempt / buildApiRoutes / API_ROUTE_PATTERNS. The
ROUTE_TABLE stays in index.ts as the canonical wire surface for
gen:docs / gen:openapi.

Test files reorganised alongside their handlers:

- handlers.test.ts (1627 lines) → handlers/agents.test.ts (533) +
  handlers/projects.test.ts (938) + handlers/brainstorm.test.ts (393)
- handlers.preview.test.ts (857 lines) → handlers/index.test.ts (27,
  isAuthExempt only) + handlers/meta.test.ts (139, /preview/config) +
  handlers/runs.preview.test.ts (767, preview login + teardown +
  proxy-preamble integration)
- handlers.burrows.test.ts and handlers.workers.test.ts move under
  handlers/ without splitting

Route handler symbol renames (listAgents → listAgentsHandler,
healthz → healthzHandler, readyz → readyzHandler, version →
versionHandler) propagate through docs/http-api.md + docs/openapi.yaml
via 'bun run gen:docs && bun run gen:openapi'.

biome.json: dropped handlers/index.ts from the
noExcessiveCognitiveComplexity override list (no longer needed after
the split); added handlers/projects.ts (runProjectTriggerHandler stays
at complexity 22 — refactoring it is out of scope).

scripts/file-size-budgets.json: removed handlers/index.ts (now 420 <
500 threshold) + handlers.test.ts + handlers.preview.test.ts; added
the three test files that landed above the 500-line default
(handlers/agents.test.ts 533, handlers/projects.test.ts 938,
handlers/runs.preview.test.ts 767). Phase 4 of pl-9088 follow-ups can
ratchet these down.

bun run check:all is green.
