+++ biome.json
@@ -123,10 +123,10 @@
 				"src/server-config/workers.ts",
 				"src/server/bridges.ts",
 				"src/server/errors.ts",
-				"src/server/handlers/index.ts",
 				"src/server/handlers/plan-runs.ts",
 				"src/server/handlers/plot-plan-runs.ts",
 				"src/server/handlers/plots.ts",
+				"src/server/handlers/projects.ts",
 				"src/server/handlers/runs.ts",
 				"src/server/integration.multi-worker.test.ts",
 				"src/server/main.ts",
