+++ scripts/file-size-budgets.json
@@ -40,9 +40,9 @@
 		"src/runs/stream.ts": 911,
 		"src/server/bridges.test.ts": 647,
 		"src/server/bridges.ts": 655,
-		"src/server/handlers.preview.test.ts": 857,
-		"src/server/handlers.test.ts": 1627,
-		"src/server/handlers/index.ts": 1194,
+		"src/server/handlers/agents.test.ts": 533,
+		"src/server/handlers/projects.test.ts": 938,
+		"src/server/handlers/runs.preview.test.ts": 767,
 		"src/server/handlers/plan-runs.test.ts": 1318,
 		"src/server/handlers/plot-plan-runs.test.ts": 640,
 		"src/server/handlers/plots.test.ts": 3726,
