+++ src/server/handlers/diagnostics.ts
@@ -1,0 +1,136 @@
+/**
+ * Diagnostics handlers — `/readyz` (warren-599c / pl-9088 step 3).
+ *
+ * Extracted from `handlers/index.ts`. The `/healthz` + `/version`
+ * inert probes live in `./meta.ts` because they don't need any of the
+ * diagnostic checks subsystem.
+ */
+
+import { DrizzleAdapter } from "../../db/repos/drizzle-adapter.ts";
+import {
+	checkBurrowPoolReachable,
+	checkBwrap,
+	checkCanopyClean,
+	checkCanopyClone,
+	checkDatabaseReachable,
+	checkPreviewAuthStrength,
+	checkPreviewMaxLive,
+	checkPreviewPortAllocator,
+	checkWarrenConfig,
+	checkWarrenConfigDeprecations,
+	type DiagnosticCheck,
+} from "../../diagnostics/checks.ts";
+import { createRunPreviewsRepo, DEFAULT_MAX_LIVE } from "../../preview/eviction.ts";
+import { DEFAULT_PREVIEW_PORT_RANGE, PreviewPortAllocator } from "../../preview/port-allocator.ts";
+import type { SpawnFn } from "../../projects/clone.ts";
+import { jsonResponse } from "../response.ts";
+import type { RouteHandler, ServerDeps } from "../types.ts";
+import { defaultSpawn } from "./index.ts";
+
+export function readyzHandler(deps: ServerDeps): RouteHandler {
+	return async () => {
+		// SpawnFn is required for the bwrap + canopy_clean probes; main.ts
+		// always wires `defaultSpawn`, but the type system keeps it
+		// optional so tests don't have to populate it. Fall back to the
+		// handler-local `defaultSpawn` to keep the contract live in tests
+		// that don't override.
+		const spawn: SpawnFn = deps.spawn ?? defaultSpawn;
+		// Canopy probes are gated on `CANOPY_REPO_URL` being configured
+		// (warren-d3e9). With no library, both probes return informational
+		// `ok: true` rather than failing — built-in agents cover the
+		// "no library" case and there's no clone to inspect.
+		const env: Readonly<Record<string, string | undefined>> =
+			deps.canopyConfig !== undefined
+				? {
+						CANOPY_REPO_URL: deps.canopyConfig.repoUrl,
+						WARREN_CANOPY_DIR: deps.canopyConfig.localDir,
+						WARREN_GIT_BINARY: deps.canopyConfig.gitBinary,
+					}
+				: {};
+
+		const checks: DiagnosticCheck[] = [];
+		checks.push(
+			await checkDatabaseReachable({ ...(deps.db !== undefined ? { db: deps.db } : {}) }),
+		);
+		checks.push(await checkBurrowPoolReachable(deps.burrowClientPool));
+		checks.push(await checkAgentsRegistered(deps));
+		checks.push(checkCanopyClone({ env }));
+		checks.push(await checkCanopyClean({ env, spawn }));
+		checks.push(await checkBwrap({ spawn }));
+		const warrenConfigProjects = (await deps.repos.projects.listAll()).map((p) => ({
+			id: p.id,
+			localPath: p.localPath,
+		}));
+		const warrenConfigArgs = {
+			projects: warrenConfigProjects,
+			...(deps.warrenConfigs !== undefined ? { cache: deps.warrenConfigs } : {}),
+		};
+		checks.push(await checkWarrenConfig(warrenConfigArgs));
+		checks.push(await checkWarrenConfigDeprecations(warrenConfigArgs));
+		checks.push(await previewPortAllocatorReadyzCheck(deps));
+		checks.push(await previewMaxLiveReadyzCheck(deps));
+		// Auth-strength probe (R-19 / SPEC §11.L, warren-8a10) reads from
+		// process.env directly: server boot already validated the token shape,
+		// so /readyz only needs to surface the strength heuristic against the
+		// live env. Tests that don't override `process.env` get the inert
+		// "preview disabled" branch.
+		checks.push(checkPreviewAuthStrength({ env: process.env }));
+
+		const allOk = checks.every((c) => c.ok);
+		return jsonResponse(allOk ? 200 : 503, {
+			ok: allOk,
+			checks,
+		});
+	};
+}
+
+async function previewPortAllocatorReadyzCheck(deps: ServerDeps): Promise<DiagnosticCheck> {
+	// Range is resolved at boot (`ServerDeps.previewPortRange`) so /readyz
+	// doesn't re-parse env per request. Tests omit deps.previewPortRange;
+	// fall back to defaults so the probe still exercises the codepath.
+	const range = deps.previewPortRange ?? DEFAULT_PREVIEW_PORT_RANGE;
+	if (deps.db === undefined) {
+		return {
+			name: "preview_port_allocator",
+			ok: true,
+			message: `no db handle wired (range ${range.start}-${range.end})`,
+		};
+	}
+	const allocator = new PreviewPortAllocator(DrizzleAdapter.for(deps.db), range);
+	return checkPreviewPortAllocator({ probe: allocator });
+}
+
+async function previewMaxLiveReadyzCheck(deps: ServerDeps): Promise<DiagnosticCheck> {
+	const maxLive = deps.previewMaxLive ?? DEFAULT_MAX_LIVE;
+	if (deps.db === undefined) {
+		return {
+			name: "preview_max_live",
+			ok: true,
+			message: `no db handle wired (cap ${maxLive})`,
+		};
+	}
+	const previews = createRunPreviewsRepo(deps.db);
+	return checkPreviewMaxLive({
+		probe: { count: () => previews.countActivePreviews() },
+		maxLive,
+	});
+}
+
+async function checkAgentsRegistered(deps: ServerDeps): Promise<DiagnosticCheck> {
+	const count = (await deps.repos.agents.listAll()).length;
+	if (count === 0) {
+		// Built-ins seed on boot (warren-d3e9), so an empty registry here
+		// means seeding itself failed — an internal problem, not an
+		// operator one. Keep the failure but reword the hint accordingly.
+		return {
+			name: "agents",
+			ok: false,
+			message: "no agents registered",
+			hint:
+				deps.canopyConfig !== undefined
+					? "POST /agents/refresh against your canopy library, or check the warren server logs for built-in seed errors"
+					: "check the warren server logs for built-in seed errors",
+		};
+	}
+	return { name: "agents", ok: true };
+}
