+++ src/server/handlers/runs.preview.test.ts
@@ -1,16 +1,15 @@
 import { afterEach, beforeEach, describe, expect, test } from "bun:test";
-import { BurrowClient, BurrowClientPool } from "../burrow-client/index.ts";
-import { type AnyWarrenDb, openDatabase, type WarrenDb } from "../db/client.ts";
-import { createRepos, type Repos } from "../db/repos/index.ts";
-import { isPostgresTestEnabled, withDb } from "../db/testing.ts";
-import { COOKIE_NAME, createPreviewAuth, type PreviewAuth } from "../preview/cookie.ts";
-import { createPreviewProxyHandler } from "../preview/proxy.ts";
-import { RunEventBroker } from "../runs/index.ts";
-import { bearerAuth } from "./auth.ts";
-import { createBridgeRegistry } from "./bridges.ts";
-import { isAuthExempt } from "./handlers/index.ts";
-import { startServer } from "./server.ts";
-import type { BridgeRegistry, ServeHandle, ServerDeps } from "./types.ts";
+import { BurrowClient, BurrowClientPool } from "../../burrow-client/index.ts";
+import { type AnyWarrenDb, openDatabase, type WarrenDb } from "../../db/client.ts";
+import { createRepos, type Repos } from "../../db/repos/index.ts";
+import { isPostgresTestEnabled, withDb } from "../../db/testing.ts";
+import { COOKIE_NAME, createPreviewAuth, type PreviewAuth } from "../../preview/cookie.ts";
+import { createPreviewProxyHandler } from "../../preview/proxy.ts";
+import { RunEventBroker } from "../../runs/index.ts";
+import { bearerAuth } from "../auth.ts";
+import { createBridgeRegistry } from "../bridges.ts";
+import { startServer } from "../server.ts";
+import type { BridgeRegistry, ServeHandle, ServerDeps } from "../types.ts";
 
 const TOKEN = "test-token-very-secret-1234567890abcdef";
 const HOST = "preview.warren.example.com";
@@ -71,28 +70,7 @@ function tcpUrl(handle: ServeHandle): string {
 	return `http://${handle.transport.hostname}:${handle.transport.port}`;
 }
 
-describe("isAuthExempt", () => {
-	test("/healthz remains auth-exempt", () => {
-		expect(isAuthExempt("/healthz")).toBe(true);
-	});
-
-	test("/version is auth-exempt (warren-6ea5)", () => {
-		expect(isAuthExempt("/version")).toBe(true);
-	});
-
-	test("/runs/<id>/preview/login is auth-exempt (SPEC §11.L)", () => {
-		expect(isAuthExempt("/runs/run_abc/preview/login")).toBe(true);
-		expect(isAuthExempt("/runs/run_abc/preview/login/")).toBe(true);
-	});
-
-	test("other /runs/* surfaces remain gated", () => {
-		expect(isAuthExempt("/runs")).toBe(false);
-		expect(isAuthExempt("/runs/run_abc")).toBe(false);
-		expect(isAuthExempt("/runs/run_abc/events")).toBe(false);
-		expect(isAuthExempt("/runs/run_abc/preview")).toBe(false);
-		expect(isAuthExempt("/runs/run_abc/preview/login/extra")).toBe(false);
-	});
-});
+/* Preview-surface tests for /runs/:id (extracted from handlers.preview.test.ts, warren-599c / pl-9088 step 3). */
 
 describe("GET /runs/:id/preview/login", () => {
 	let db: WarrenDb;
@@ -398,74 +376,6 @@ describe("GET /runs/:id/preview/login", () => {
 			);
 			expect(res.status).toBe(302);
 		});
-	});
-});
-
-describe("GET /preview/config (warren-016d)", () => {
-	let db: WarrenDb;
-	let repos: Repos;
-	let handle: ServeHandle | null = null;
-
-	beforeEach(async () => {
-		db = await openDatabase({ path: ":memory:" });
-		repos = createRepos(db);
-	});
-
-	afterEach(async () => {
-		if (handle) {
-			await handle.stop();
-			handle = null;
-		}
-		await db.close();
-	});
-
-	test("returns mode + host in subdomain mode", async () => {
-		const previewAuth = createPreviewAuth(TOKEN, {
-			scope: { mode: "subdomain", cookieDomain: `.${HOST}` },
-			secure: false,
-		});
-		const { deps } = await depsFor(repos, previewAuth);
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: bearerAuth(TOKEN),
-			logger: silentLogger,
-		});
-		const res = await fetch(`${tcpUrl(handle)}/preview/config`, {
-			headers: { authorization: `Bearer ${TOKEN}` },
-		});
-		expect(res.status).toBe(200);
-		const body = (await res.json()) as { mode: string; host: string | null };
-		expect(body.mode).toBe("subdomain");
-		expect(body.host).toBe(HOST);
-	});
-
-	test("returns mode + null host in path mode without WARREN_PREVIEW_HOST", async () => {
-		const previewAuth = createPreviewAuth(TOKEN, { scope: { mode: "path" }, secure: false });
-		const { deps } = await depsFor(repos, previewAuth, undefined, "path");
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: bearerAuth(TOKEN),
-			logger: silentLogger,
-		});
-		const res = await fetch(`${tcpUrl(handle)}/preview/config`, {
-			headers: { authorization: `Bearer ${TOKEN}` },
-		});
-		expect(res.status).toBe(200);
-		const body = (await res.json()) as { mode: string; host: string | null };
-		expect(body.mode).toBe("path");
-		expect(body.host).toBeNull();
-	});
-
-	test("401 without a bearer token (gated like every non-login preview surface)", async () => {
-		const previewAuth = createPreviewAuth(TOKEN, { scope: { mode: "path" }, secure: false });
-		const { deps } = await depsFor(repos, previewAuth, undefined, "path");
-		handle = startServer(deps, {
-			transport: { kind: "tcp", hostname: "127.0.0.1", port: 0 },
-			auth: bearerAuth(TOKEN),
-			logger: silentLogger,
-		});
-		const res = await fetch(`${tcpUrl(handle)}/preview/config`);
-		expect(res.status).toBe(401);
 	});
 });
